+++ src/main/java/messagesApp/message/MessageDao.java
@@ -4,24 +4,26 @@ import messagesApp.global.Global;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
 
 public class MessageDao {
-    public static boolean createMessage (Message message) {
+    public static Map<Integer, Boolean> createMessage (Message message) {
         PreparedStatement ps;
-        boolean successCode;
+        Map<Integer, Boolean> result = new HashMap<Integer, Boolean>();
         try {
             String query = "INSERT INTO messages (message, author) VALUES (?, ?)";
             ps = Global.psqlConnection.prepareStatement(query);
             ps.setString(1, message.getMessage());
             ps.setString(2, message.getAuthor());
             ps.executeUpdate();
-            successCode = true;
+            result.put(1, true);
         } catch (SQLException e) {
             e.printStackTrace();
-            successCode = false;
+            result.put(1, false);
         }
 
-        return successCode;
+        return result;
     }
 
     public static void readMessage () {