+++ src/main/java/messagesApp/message/MessageService.java
@@ -1,9 +1,10 @@
 package messagesApp.message;
 
+import java.util.Map;
 import java.util.Scanner;
 
 public class MessageService {
-    public static void createMessage () {
+    public static String createMessage () {
         Scanner sc = new Scanner(System.in);
         System.out.print("Write a message: ");
         String messageContent = sc.nextLine();
@@ -11,8 +12,10 @@ public class MessageService {
         String messageAuthor = sc.nextLine();
 
         Message newMessage = new Message(messageContent, messageAuthor);
-        boolean success = MessageDao.createMessage(newMessage);
-        if (success) System.out.println("\nMessage created successfully!\n\n");
+        Map<Integer, Boolean> result = MessageDao.createMessage(newMessage);
+        if (result.get(1)) return "\nMessage created successfully!\n\n";
+
+        return "\nThere was an error trying to create the requested message!";
     }
 
     public static void listMessages () {